Fix error reported by recursive_dir_it for dangling symlinks.
This only affects POSIX platforms not supporting openat & co.

When the underlying directory iterator does not produce symlink status
during iteration, and the iterator points to a dangling symlink, the
recursive_directory_iterator increment would attempt to refresh file
statuses in the directory_entry. This would fail because the refresh
would query status in addition to symlink_status. This error was
wrongly reported to the caller if following symlinks was disabled for
the recursive iterator.

Fix this by only querying symlink_status. status is checked later,
if the iterator is configured to follow symlinks.