Rearranged code of pushing a directory in recursive_dir_it increment.
Avoid indirectly calling refresh on the directory entry that we are
checking for whether we can descend into it. Calling refresh is
wasteful as it always queries both symlink_status and status and we
already have the actual symlink_status by the time we need status.

Rearranged code to group POSIX branches closer together and slightly
reduce code duplication.